Version 3.05   * Scheduler: expose application ID   * Add example script roi-iter.py to dynamically select ROI based on SimMarkers   * CPI stacks: --aggregate and --partial support, fixes for heterogeneous configurations   * Traces: support for 32-bit executables   * Build fixes for older Linux versions

+++ pin/codecache_trace.cc
@@ -114,7 +114,7 @@ void initCodeCacheTracing()
 
    next_callback = SubsecondTime::Zero();
 
-   Sim()->getHooksManager()->registerHook(HookType::HOOK_PERIODIC, (HooksManager::HookCallbackFunc)codeCachePeriodicCallback, NULL);
+   Sim()->getHooksManager()->registerHook(HookType::HOOK_PERIODIC, (HooksManager::HookCallbackFunc)codeCachePeriodicCallback, (UInt64)NULL);
 
    //CODECACHE_AddCacheInitFunction(cacheInit, NULL);
    CODECACHE_AddCacheBlockFunction(newCacheBlock, NULL);
